requirements.txt
# Backend dependencies
strawberry-graphql==0.183.6
fastapi==0.95.1
uvicorn==0.22.0
pandas==2.0.1
numpy==1.24.3
python-dotenv==1.0.0

# Frontend dependencies
streamlit==1.22.0
gql==3.4.1
requests==2.30.0
plotly==5.14.1
matplotlib==3.7.1

# Development dependencies
pytest==7.3.1
black==23.3.0
flake8==6.0.0